+++ src/simsopt/field/biotsavart.py
@@ -6,63 +6,6 @@ from simsopt._core.graph_optimizable import Optimizable
 from simsopt._core.derivative import Derivative
 
 
-class Coil(sopp.Coil, Optimizable):
-    """
-    A :mod:`Coil` combines a :mod:`Curve` and a :mod:`Current` and is used as input
-    for the mod:`BiotSavart` field.
-    """
-
-    def __init__(self, curve, current):
-        self.__curve = curve
-        self.__current = current
-        sopp.Coil.__init__(self, curve, current)
-        Optimizable.__init__(self, x0=np.asarray([]), depends_on=[curve, current])
-
-    def vjp(self, v_gamma, v_gammadash, v_current):
-        return self.curve.dgamma_by_dcoeff_vjp(v_gamma) \
-            + self.curve.dgammadash_by_dcoeff_vjp(v_gammadash) \
-            + self.current.vjp(v_current)
-
-
-class Current(sopp.Current, Optimizable):
-    """
-    An optimizable object that wraps around a single scalar degree of freedom
-    and represents a coil.
-    """
-
-    def __init__(self, current):
-        sopp.Current.__init__(self, current)
-        Optimizable.__init__(self, external_dof_setter=Current.set_dofs,
-                             x0=self.get_dofs())
-
-    def set_dofs(self, dofs):
-        sopp.Current.set_dofs(self, dofs)
-
-    def vjp(self, v_current):
-        return Derivative({self: v_current})
-
-    def __neg__(self):
-        return ScaledCurrent(self, -1.)
-
-
-class ScaledCurrent(sopp.ScaledCurrent, Optimizable):
-    """
-    Scales :mod:`Current` by a factor. To be used for example to flip currents
-    for stellarator symmetric coils.
-    """
-
-    def __init__(self, basecurrent, scale):
-        self.__basecurrent = basecurrent
-        sopp.ScaledCurrent.__init__(self, basecurrent, scale)
-        Optimizable.__init__(self, x0=np.asarray([]), depends_on=[basecurrent])
-
-    def vjp(self, v_current):
-        return self.__basecurrent.vjp(self.scale * v_current)
-
-    def __neg__(self):
-        return ScaledCurrent(self, -1.)
-
-
 class BiotSavart(sopp.BiotSavart, MagneticField):
     r"""
     Computes the MagneticField induced by a list of closed curves :math:`\Gamma_k` with electric currents :math:`I_k`.
