+++ src/simsopt/geo/curve.py
@@ -806,3 +806,34 @@ def curves_to_vtk(curves, filename):
     ppl = np.asarray([c.gamma().shape[0] for c in curves])
     data = np.concatenate([i*np.ones((curves[i].gamma().shape[0], )) for i in range(len(curves))])
     polyLinesToVTK(filename, x, y, z, pointsPerLine=ppl, pointData={'idx': data})
+
+
+def create_equally_spaced_curves(ncurves, nfp, stellsym, R0=1.0, R1=0.5, order=6, PPP=15):
+    """
+    Create ``ncurves`` curves that will result in circular equally spaced coils
+    (major radius ``R0`` and minor radius ``R1``)  after applying
+    ``coils_via_symmetries``.
+
+    Usage example: create 4 base curves, which are then rotated 3 times and
+    flipped for stellarator symmetry:
+
+    .. code-block::
+
+        base_curves = create_equally_spaced_curves(4, 3, stellsym=True)
+        base_currents = [Current(1e5) for c in base_curves]
+        coils = coils_via_symmetries(base_curves, base_currents, 3, stellsym=True)
+
+    """
+    curves = []
+    from simsopt.geo.curvexyzfourier import CurveXYZFourier
+    for i in range(ncurves):
+        curve = CurveXYZFourier(order*PPP, order)
+        d = curve.x
+        d[0] = R0
+        d[1] = R1
+        d[2*(2*order+1)+2] = R1
+        curve.x = d
+        angle = (i+0.5)*(2*np.pi)/((1+int(stellsym))*nfp*ncurves)
+        curve = RotatedCurve(curve, angle, False)
+        curves.append(curve)
+    return curves
